requirements.txt
# Vwap - Vegan Recipe Swap Backend Dependencies

# Core Flask framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Authentication and Security
Werkzeug==3.0.1
bcrypt==4.1.1
PyJWT==2.8.0

# Environment and Configuration
python-dotenv==1.0.0

# File Upload and Image Processing
Pillow==10.1.0
Flask-Uploads==0.2.1

# Email
Flask-Mail==0.9.1

# Caching and Sessions
Flask-Session==0.5.0
redis==5.0.1

# API Documentation
Flask-RESTful==0.3.10
flask-restx==1.3.0
marshmallow==3.20.2

# Validation
cerberus==1.3.5
email-validator==2.1.0

# HTTP Requests
requests==2.31.0

# Date and Time
python-dateutil==2.8.2

# Utilities
click==8.1.7
gunicorn==21.2.0

# Development and Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Monitoring and Logging
sentry-sdk[flask]==1.38.0

# Production
gevent==23.9.1
